CVE-2015-1774-hwpreader-check-reads
Gbp-Pq: Name CVE-2015-1774-hwpreader-check-reads.patch

+++ hwpfilter/source/htags.h
@@ -67,7 +67,7 @@ struct HyperText
 struct OlePicture
 {
     int   size;
-    ulong signature;
+    uint signature;
 #ifdef WIN32
      IStorage *pis;
 #else
